Правки редактирование комментариев

+++ src/BlogEngine.Web/Controllers/CommentController.cs
@@ -1,5 +1,6 @@
 ﻿using BlogEngine.Core.Enums;
 using BlogEngine.Service.Attributes;
+using BlogEngine.Service.Exceptions;
 using BlogEngine.Service.Models;
 using BlogEngine.Service.Services;
 using Microsoft.AspNetCore.Mvc;
@@ -28,9 +29,14 @@ public class CommentController : CRUDControllerBase<CommentContract>
     /// <summary>
     ///     Обновить
     /// </summary>
-    [Allowed(UserRole.ADMIN)]
+    [Allowed(UserRole.ADMIN, UserRole.USER)]
     public override async Task<IActionResult> UpdateAsync([FromBody] CommentContract item)
     {
+        if (item.Author.Nickname != Nickname)
+        {
+            throw new AppException("You're not allowed to edit this comment", System.Net.HttpStatusCode.Forbidden);
+        }
+
         var result = await _service.UpdateAsync(item);
         return Ok();
     }
@@ -38,9 +44,16 @@ public class CommentController : CRUDControllerBase<CommentContract>
     /// <summary>
     ///     Удалить по id
     /// </summary>
-    [Allowed(UserRole.ADMIN)]
+    [Allowed(UserRole.ADMIN, UserRole.USER)]
     public override async Task<IActionResult> DeleteAsync([FromBody] long id)
     {
+        var item = await _service.GetByIdAsync(id);
+
+        if (item.Author.Nickname != Nickname && Role != UserRole.ADMIN)
+        {
+            throw new AppException("You're not allowed to delete this comment", System.Net.HttpStatusCode.Forbidden);
+        }
+
         var result = await _service.DeleteAsync(id);
         return Ok();
     }
